move vaos and vass calendar to shared
