+++ src/applications/vaos/new-appointment/components/DateTimeSelectPage/index.jsx
@@ -9,6 +9,7 @@ import {
 } from 'date-fns';
 import { formatInTimeZone } from 'date-fns-tz';
 import { scrollToFirstError } from 'platform/utilities/scroll';
+import CalendarWidget from 'platform/shared/calendar/CalendarWidget';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
@@ -19,7 +20,6 @@ import {
   getUpcomingAppointmentListInfo,
   selectUpcomingAppointments,
 } from '../../../appointment-list/redux/selectors';
-import CalendarWidget from '../../../components/calendar/CalendarWidget';
 import FormButtons from '../../../components/FormButtons';
 import InfoAlert from '../../../components/InfoAlert';
 import NewTabAnchor from '../../../components/NewTabAnchor';
