+++ src/applications/vaos/referral-appointments/components/DateAndTimeContent.jsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import CalendarWidget from '../../components/calendar/CalendarWidget';
+import CalendarWidget from 'platform/shared/calendar/CalendarWidget';
 import { setSelectedSlotStartTime } from '../redux/actions';
 import FormButtons from '../../components/FormButtons';
 import { routeToNextReferralPage, routeToPreviousReferralPage } from '../flow';
