--- src/applications/vass/components/calendar/CalendarContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const CalendarContext = createContext(null);
