+++ src/applications/vass/pages/DateTimeSelection.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useCallback, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom-v5-compat';
 import { focusElement } from 'platform/utilities/ui/focus';
+import CalendarWidget from 'platform/shared/calendar/CalendarWidget';
 import Wrapper from '../layout/Wrapper';
 import { usePersistentSelections } from '../hooks/usePersistentSelections';
 import { setSelectedDate, selectSelectedDate } from '../redux/slices/formSlice';
@@ -9,8 +10,6 @@ import { useGetAppointmentAvailabilityQuery } from '../redux/api/vassApi';
 // TODO: remove this once we have a real UUID
 import { UUID } from '../services/mocks/utils/formData';
 
-// TODO: make this component a shared component
-import CalendarWidget from '../components/calendar/CalendarWidget';
 import { mapAppointmentAvailabilityToSlots } from '../utils/slots';
 import { getTimezoneDescByTimeZoneString } from '../utils/timezone';
