+++ src/platform/shared/calendar/CalendarNavigation.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable @department-of-veterans-affairs/prefer-button-component */
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
