+++ src/platform/shared/calendar/CalendarOptionsSlots.jsx
