+++ src/platform/shared/calendar/utils/utils.js
@@ -13,7 +13,7 @@ import {
   isWeekend,
   startOfDay,
 } from 'date-fns';
-import { APPOINTMENT_STATUS, DATE_FORMATS } from './constants';
+import { APPOINTMENT_STATUS, DATE_FORMATS } from '../constants';
 /**
  * Ensures the input is a Date object, converting from string if needed
  * @param {Date|string} value - Date object or date string
